app/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncpg==0.30.0
bcrypt==4.2.1
black==24.10.0
build==1.2.2.post1
CacheControl==0.14.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.7
crashtest==0.4.1
cryptography==43.0.3
distlib==0.3.9
dnspython==2.7.0
dulwich==0.21.7
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.3
fastjsonschema==2.20.0
filelock==3.16.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
installer==0.7.0
jaraco.classes==3.4.0
keyring==24.3.1
Mako==1.3.6
MarkupSafe==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
mypy-extensions==1.0.0
packaging==24.1
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pkginfo==1.11.2
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.4
poetry-core==1.9.1
poetry-plugin-export==1.8.0
psycopg==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.6.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.9.0
pyproject_hooks==1.2.0
pytest==8.3.3
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.16
RapidFuzz==3.10.0
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
setuptools==75.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.0
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
virtualenv==20.27.0
xattr==1.1.0
